Added two new CLI options. --run-pre-wash, --run-post-wash.
This now enables utilising the $plugin->sortorder to determine a two stash washing process.

Sub plugins that have a sort order of <200 will be run in the --run-pre-wash.
Sub plugins that have a sort order of >=200 will be run in the --run-post-wash.

When only the --run option is specified, all sub plugins will run.